+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
-## v. 0.1.3
+## v. 0.1.4
 
-- demo : audio now play just if recording is not aborted
+- event names 
+
+  recording* -> speech*
+
+- demo 
+
+  - colors: a bit clearer 
+  - audio now play just if recording is not aborted